requirements.txt
streamlit
pandas
chromadb-client
sentence-transformers
google-generativeai
langchain-core
torch